# docs/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OracleMonitor API",
    "version": "1.0.0",
    "description": "Universal Oracle Monitoring Platform API",
    "contact": {
      "name": "API Support",
      "email": "api@oracle-monitor.foresight.build"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/oracle/config-history/:id": {
      "get": {
        "summary": "GET /oracle/config-history/:id",
        "description": "Endpoint for GET requests to /oracle/config-history/:id",
        "tags": ["Config"],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "POST /oracle/config-history/:id",
        "description": "Endpoint for POST requests to /oracle/config-history/:id",
        "tags": ["Config"],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/oracle/config-history/stats": {
      "get": {
        "summary": "GET /oracle/config-history/stats",
        "description": "Endpoint for GET requests to /oracle/config-history/stats",
        "tags": ["Config"],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/oracle/uma-users/[address]/assertions": {
      "get": {
        "summary": "GET /oracle/uma-users/[address]/assertions",
        "description": "Endpoint for GET requests to /oracle/uma-users/[address]/assertions",
        "tags": ["UMA"],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/oracle/uma-users/[address]/stats": {
      "get": {
        "summary": "GET /oracle/uma-users/[address]/stats",
        "description": "Endpoint for GET requests to /oracle/uma-users/[address]/stats",
        "tags": ["UMA"],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}